+++ sdk/netapp/Azure.ResourceManager.NetApp/src/Custom/Models/ElasticBackupSnapshotUsage.ElasticSerializationCompatibility.cs
@@ -1,0 +1,15 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+#nullable disable
+
+#pragma warning disable CS1591
+
+namespace Azure.ResourceManager.NetApp.Models
+{
+    public readonly partial struct ElasticBackupSnapshotUsage
+    {
+        public static ElasticBackupSnapshotUsage CreateNewSnapshot { get; } = new ElasticBackupSnapshotUsage("CreateNewSnapshot");
+        public static ElasticBackupSnapshotUsage UseExistingSnapshot { get; } = new ElasticBackupSnapshotUsage("UseExistingSnapshot");
+    }
+}
