--- sdk/netapp/Azure.ResourceManager.NetApp/src/Custom/Models/ElasticMoreSerializationCompatibility.cs
@@ -1,16 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-#nullable disable
-
-#pragma warning disable SA1402
-#pragma warning disable SA1649
-#pragma warning disable CS1591
-
-using System.ClientModel.Primitives;
-using System.Text.Json;
-using Azure;
-
-namespace Azure.ResourceManager.NetApp.Models
-{
-}
